refactor(bridge,events): table-driven adapters and dispatchers
Replaces the two large hand-rolled switches with mapped-type tables that
the type checker validates exhaustively.

Bridge layer
- New `Bridge/schema.ts` exposes `ADAPTERS satisfies AdapterMap`, where
  `AdapterMap = { [K in WhatsAppEvent['type']]: (data: BridgeData<K>) => CanonicalEvent | null }`.
  Each entry receives the bridge's typed `data` slot — TS catches missing
  variants at compile time. Adding a new bridge event without an adapter
  is a build error, not a silent drop.
- `KNOWN_BRIDGE_EVENT_TYPES` now derives from `Object.keys(ADAPTERS)` —
  one source of truth, no risk of the set drifting from the table.
- `adapt.ts` shrinks to a thin shim re-exporting through `schema.ts` so
  call sites and the dts-drift test remain stable.

Dispatch layer
- `events.ts` switch replaced by `DISPATCHERS satisfies DispatcherMap`,
  keyed by `CanonicalEvent['type']`. Each handler receives the narrow
  event payload (`Extract<CanonicalEvent, { type: K }>`) without `as`
  casts. Forgetting a handler is a compile error.
- Centralized the `Partial<ConnectionState>` cast in
  `emitConnectionUpdate` so individual dispatchers stop carrying it.
- The handler now wraps each dispatch in try/catch — one bad event no
  longer poisons the pipeline.

Type safety net
- `as never` / `as unknown as Record<…>` reduced to 8 documented
  boundary casts (each at a bridge-untyped→canonical edge), down from
  ~20 spread through both files.
- Remaining casts are flagged with comments explaining why narrowing
  fails (mostly bridge sub-types like PinAction that aren't separately
  exported by the bridge .d.ts and resolve to `any`).

All 161 existing tests pass; no behavioral change.
